Create CollapsedStoryInfo component
