+++ app/assets/javascripts/components/story/CollapsedStory/CollapsedStory.js
@@ -5,12 +5,12 @@ import StoryPopover from '../StoryPopover';
 import StoryDescriptionIcon from '../StoryDescriptionIcon';
 import CollapsedStoryEstimate from './CollapsedStoryEstimate';
 import CollapsedStoryStateActions from './CollapsedStoryStateActions';
+import CollapsedStoryInfo from './CollapsedStoryInfo';
 
 import {
   classIconRule,
   iconRule,
   isStoryNotEstimated,
-  labelSplit,
   isRelease,
 } from '../../../rules/story';
 
@@ -22,62 +22,6 @@ const StoryIcon = ({ storyType }) => (
 
 StoryIcon.propTypes = {
   storyType: PropTypes.string.isRequired,
-};
-
-const StoryLabel = ({ label }) => (
-  <a href="#" className="Story__label" title={label}>{label}</a>
-);
-
-StoryLabel.propTypes = {
-  label: PropTypes.string.isRequired,
-};
-
-const StoryLabels = ({ labels }) => {
-  if (!labels) {
-    return null
-  }
-
-  return (
-    <span className='Story__labels'>
-      {labelSplit(labels).map(label => (
-        <StoryLabel key={label} label={label} />
-      ))}
-    </span>
-  );
-};
-
-export const StoryTitle = ({ title, ownedByInitials, ownedByName }) => (
-  <div className="Story__title">
-    {title}
-    <abbr
-      className="Story__initials"
-      title={ownedByName}
-    >
-      {ownedByInitials}
-    </abbr>
-  </div>
-);
-
-StoryTitle.propTypes = {
-  title: PropTypes.string.isRequired,
-  ownedByName: PropTypes.string,
-  ownedByInitials: PropTypes.string,
-};
-
-const StoryInfo = ({ title, labels, ownedByName, ownedByInitials }) => (
-  <div className="Story__info">
-    <StoryLabels labels={labels} />
-    <StoryTitle title={title} ownedByInitials={ownedByInitials} ownedByName={ownedByName} />
-  </div>
-);
-
-StoryInfo.propTypes = {
-  title: PropTypes.string.isRequired,
-  labels: PropTypes.string,
-};
-
-StoryInfo.defaultProps = {
-  labels: '',
 };
 
 const classNameStory = (storyType, estimate) => classname(
@@ -106,7 +50,7 @@ const CollapsedStory = ({ story }) => (
       </div>
     </StoryPopover>
 
-    <StoryInfo
+    <CollapsedStoryInfo
       title={story.title}
       labels={story.labels}
       ownedByInitials={story.ownedByInitials}
